Inline SearchSelect and IconField helpers into their components.
Remove separate utility files and keep option-mapping logic colocated with each field.
