+++ app/react/V2/Components/Metadata/EntityEditor/Components/IconField.tsx
@@ -1,13 +1,76 @@
 import React from 'react';
 import { Controller, FieldValues, useFormContext } from 'react-hook-form';
+import { iconNames } from '#UI/Icon/library.js';
+import { CountryList } from '#app/UI/index.js';
+import { Icon } from '#UI/Icon/Icon.jsx';
 import { Translate } from '#app/I18N/index.js';
+import { CountryFlag } from '#V2/Components/CustomIcons/CoutryFlags.js';
 import { Checkbox, SearchSelect } from '#V2/Components/Forms/index.js';
-import { iconSelectGroups } from '../functions/iconSelectOptions.js';
-import {
-  iconFromSelection,
-  selectionFromIcon,
-  type EntityIcon,
-} from '../functions/iconSelectUtils.js';
+
+type EntityIcon = {
+  _id: string | null;
+  type: string;
+  label: string;
+};
+
+const EMPTY_ICON: EntityIcon = { _id: null, type: 'Empty', label: '' };
+
+const hasEntityIcon = (icon?: EntityIcon | null): boolean =>
+  Boolean(icon?._id && icon.type !== 'Empty');
+
+const selectionFromIcon = (icon?: EntityIcon | null): string => {
+  if (!hasEntityIcon(icon)) {
+    return '';
+  }
+
+  return `${icon!.type}:${icon!._id}`;
+};
+
+const iconFromSelection = (value: string): EntityIcon => {
+  if (!value) {
+    return EMPTY_ICON;
+  }
+
+  const separatorIndex = value.indexOf(':');
+  const type = value.slice(0, separatorIndex);
+  const id = value.slice(separatorIndex + 1);
+
+  if (type === 'Flags') {
+    const country = CountryList.get(id);
+    return {
+      _id: id,
+      type,
+      label: country?.label ?? id,
+    };
+  }
+
+  return {
+    _id: id,
+    type,
+    label: id,
+  };
+};
+
+const iconSelectGroups = [
+  {
+    label: 'Icons',
+    options: iconNames.map(name => ({
+      value: `Icons:${name}`,
+      searchLabel: name,
+      label: name,
+      prefix: <Icon icon={name} />,
+    })),
+  },
+  {
+    label: 'Flags',
+    options: Array.from(CountryList).map(([, country]) => ({
+      value: `Flags:${country.cca3}`,
+      searchLabel: country.label,
+      label: country.label,
+      prefix: <CountryFlag id={country.cca3} />,
+    })),
+  },
+];
 
 type IconFieldFormValues = FieldValues & {
   showIcon: boolean;
@@ -69,4 +132,5 @@ const IconField = ({ disabled = false }: IconFieldProps) => {
   );
 };
 
-export { IconField };
+export type { EntityIcon };
+export { IconField, EMPTY_ICON, hasEntityIcon };
